fix: useMessagesListSubscription hook
