+++ packages/components/modules/messages/native/index.ts
@@ -1,0 +1,2 @@
+// export native messages components
+export * from './graphql/subscriptions/useMessagesListSubscription'
